chg: Fixed shell script indent/beautified them

+++ linux/contrib/html/make_index.sh
@@ -11,12 +11,6 @@
 # Revision 4.18  2010-04-27 09:31:02  ralproth
 # cfg4.70-24165: fixes for cfg2html_solaris 1.6.4
 #
-# Revision 4.17  2009-12-16 08:33:55  ralproth
-# cfg4.60-23619: Changed the copyright notice to y2k10
-#
-# Revision 4.16  2009/03/11 15:00:39  ralproth
-# cfg4.23-22229: Forced commit by ./MakeRelease.sh for binary code distribution.
-#
 # Revision 4.15  2009/03/09 14:02:21  ralproth
 # cfg4.23-22229: Enhancements for Brocade and Itanium
 #
@@ -36,8 +30,8 @@
 # Enhanced cron collector
 #
 #######################################################################
-# (c) 1999-2010 by cfg2html@hotmail.com, All Rights Reserved, Freeware
-# http://come.to/rose_swe and come.to/cfg2html
+# (c) 1999-2018 by cfg2html@hotmail.com, All Rights Reserved, Freeware
+# http://rose.rult.at/
 #######################################################################
 # Simply run this shell script in the directory, where your cfg2html
 # files are stored. After running this script, load the allhosts.htm
@@ -68,33 +62,33 @@ EOF
 
 for i in `(find . -name "*.htm*" -print| grep -v -E './index.htm|./allhosts.htm|./info.htm' | sort -u )`  ## this find doesn't support spaces in the filename!
 do
-    # echo "Host= ["$i"]"
-    if (grep -e cfg2Html -e "cfg2html/HPUX" -e "cfg2html/HP-UX" -e Cfg2Html \
-        -e cfg2html-linux -e "cfg2html-brocade" -e "cfg2html/SUN" -e cfg2html_solaris "$i">/dev/null) ;
+  # echo "Host= ["$i"]"
+  if (grep -e cfg2Html -e "cfg2html/HPUX" -e "cfg2html/HP-UX" -e Cfg2Html \
+    -e cfg2html-linux -e "cfg2html-brocade" -e "cfg2html/SUN" -e cfg2html_solaris "$i">/dev/null) ;
+  then
+    typ2=""
+    ########### 0.54 changes ##########
+    typ=`grep "HP-UX " "$i" | grep "uname -a" | head -1 | awk '{print $5"-"$7}'`
+    if [ -z "$typ" ]
     then
-	typ2=""
-	########### 0.54 changes ##########
-	typ=`grep "HP-UX " "$i" | grep "uname -a" | head -1 | awk '{print $5"-"$7}'`
-	if [ -z "$typ" ]
-	then
-		typ=`grep "kernel.osrelease" "$i"|grep "= "|head -1 | awk -F"= " '{print $2;}'	# cut -f3 -d" "`
-		[ -z "$typ" ] && typ=$(grep "<B>SunOS" "$i"| head -1) ## <PRE><B>SunOS 5.9</B></PRE>
-                [ -z "$typ" ] && typ=$(grep "Kernel version: SunOS" "$i"| head -1|awk '{print $3,$4,$5,$6;}') ## Kernel version: SunOS 5.8 Generic 108528-24 Sep 2003</PRE>
-	else
-		typ2=`grep 9000 "$i" | grep -E '0/8|Itanium'| head -1 | sed 's+^.*9000/\(.*\)$+\1+g' `
-	fi
-	echo "Added host: $i ($typ)"
-	echo "<A HREF=\"$i\" TARGET=\"info\"><b>$i</b>" >> $OUT
-	echo "</A> " >> $OUT
-
-	if [ "$1" = "-v" ] ;
-	then
-		echo "($typ $typ2)" >> $OUT
-	fi
-	echo "<br>" >> $OUT
+      typ=`grep "kernel.osrelease" "$i"|grep "= "|head -1 | awk -F"= " '{print $2;}'	# cut -f3 -d" "`
+      [ -z "$typ" ] && typ=$(grep "<B>SunOS" "$i"| head -1) ## <PRE><B>SunOS 5.9</B></PRE>
+      [ -z "$typ" ] && typ=$(grep "Kernel version: SunOS" "$i"| head -1|awk '{print $3,$4,$5,$6;}') ## Kernel version: SunOS 5.8 Generic 108528-24 Sep 2003</PRE>
     else
-        echo "Skipping: " $i
+      typ2=`grep 9000 "$i" | grep -E '0/8|Itanium'| head -1 | sed 's+^.*9000/\(.*\)$+\1+g' `
     fi
+    echo "Added host: $i ($typ)"
+    echo "<A HREF=\"$i\" TARGET=\"info\"><b>$i</b>" >> $OUT
+    echo "</A> " >> $OUT
+
+    if [ "$1" = "-v" ] ;
+    then
+      echo "($typ $typ2)" >> $OUT
+    fi
+    echo "<br>" >> $OUT
+  else
+    echo "Skipping: " $i
+  fi
 done
 
 echo "<p><hr><p>Created: `date +%x-%X`">>$OUT
